This pull request includes several improvements and refactors across the
codebase, focusing on standardizing environment variable naming,
updating factory definitions to use the global `fake()` helper, and
making minor code cleanups and corrections. Additionally, some model
attribute accessors and mutators were reorganized for consistency, and
an unused Dockerfile was removed.

**Environment variable standardization:**

* Updated `.env.example`, `.env.pipeline`, and `.env.testing.example` to
consistently use `HE4RT_SEASON_ID` and `HE4RT_BOT_SECRET` instead of
`SEASON_ID` and `BOT_SECRET`.

**Factory and faker usage improvements:**

* Updated model factories (e.g., `CharacterFactory`, `WalletFactory`,
and `BadgeFactory`) to use the global `fake()` helper instead of
`$this->faker`, improving readability and aligning with Laravel best
practices.
* Fixed import order in factory files for consistency.

**Model accessor/mutator organization:**

* Moved `getRankingAttribute` and `getLevelAttribute` methods in
`Character` and `getMessagesCountAttribute` in `Provider` to the end of
their respective model classes for better organization and consistency.
* Refactored casts definition in `Meeting` model to use a method instead
of a property.
* Reordered factory method in `MeetingType` model for clarity.

**Code cleanup and minor fixes:**

* Removed the unused `Dockerfile` from the repository.
* Simplified constructors and class definitions for brevity in several
domain/entity classes.
* Changed `StringValueObject` from a final class to an abstract class to
allow extension.
* Fixed a minor typo in `DomainManager` (`\FilesystemIterator` to
`FilesystemIterator`).
* Corrected import order in `FindProfile.php`.
* Reordered method definitions in `MeetingRequest` for consistency.

+++ Heart/Character/Domain/Entities/LevelEntity.php
@@ -40,19 +40,19 @@ final class LevelEntity
         $memberStatusMultiplier = $isSupporter ? 0.25 : 0.4;
         $messagePound = ($messageLength / $averageMessageLength);
 
-        $experienceObtained = ($messagePound / ($this->checkLevelEntity($this->level) * $memberStatusMultiplier) * 20);
+        $experienceObtained = (int) ($messagePound / ($this->checkLevelEntity($this->level) * $memberStatusMultiplier) * 20);
         $this->addExperience($experienceObtained);
 
-        return (int) $experienceObtained;
+        return $experienceObtained;
     }
 
     public function generateVoiceExperience(VoiceStatesEnum $states, bool $isSupporter = false): int
     {
         $memberStatusMultiplier = $isSupporter ? 0.25 : 0.4;
-        $experienceObtained = ($states->getExperienceMultiplier() / ($this->level * $memberStatusMultiplier) * 20);
+        $experienceObtained = (int) ($states->getExperienceMultiplier() / ($this->level * $memberStatusMultiplier) * 20);
         $this->addExperience($experienceObtained);
 
-        return (int) $experienceObtained;
+        return $experienceObtained;
     }
 
     public function getExperience(): int
